requirements.txt
gradio
openai
python-dotenv
requests
pathlib
